Linked validator to Rabbit MQ queue
Created a rabbit package and added listener.
Created model for file_upload and error_report table.
Created repositories for FileUpload and ErrorReport Model.
Created services for file upload and error reports and creating methods to save and update this entities in the database.
Linking the validator to save errors and update validator time and file status.